+++ packages/Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
-  "soroban/*",
+  "ContractsScaffoldRust/*",
 ]
 
 [workspace.dependencies]
