+++ packages/ContractsScaffoldRust/simple-escrow/src/datatype.rs
@@ -1,0 +1,12 @@
+use soroban_sdk::contracttype;
+
+#[derive(Clone)]
+#[contracttype]
+pub enum DataKey {
+    Buyer,
+    Seller,
+    Escrow,
+    Amount,
+    Token,
+    Status,
+}
