+++ packages/ContractsScaffoldRust/simple-escrow/src/lib.rs
@@ -1,0 +1,4 @@
+#![no_std]
+mod datatype;
+mod simple_escrow_contract;
+mod test;
